Split sidebar normalization and utils and import via shadowing
