+++ plugins/gatsby-remark-dvc-linker/apiLinker.js
@@ -1,7 +1,10 @@
 /* eslint-env node */
 
 const { createLinkNode } = require('./helpers')
-const { getItemByPath } = require('../../src/utils/shared/sidebar')
+const {
+  getItemByPath
+} = require('../gatsby-theme-iterative-docs/sidebar-helpers')
+const sidebar = require('../gatsby-theme-iterative-docs/resolve-sidebar')
 
 const DVC_API_REGEXP = /dvc.api([a-z-._]*\(\)$)?/
 const METHOD_REGEXP = /^[a-z-._]*\(\)$/
@@ -25,7 +28,7 @@ module.exports = astNode => {
       url = `${API_ROOT}${method}`
     }
 
-    const isMethodPageExists = getItemByPath(url)
+    const isMethodPageExists = getItemByPath(sidebar, url)
     if (isMethodPageExists) {
       createLinkNode(url, astNode)
     }
