+++ plugins/gatsby-theme-iterative-docs/constants.js
@@ -1,0 +1,5 @@
+module.exports = {
+  PATH_ROOT: '/doc',
+  FILE_ROOT: '/docs/',
+  FILE_EXTENSION: '.md'
+}
