+++ plugins/gatsby-theme-iterative-docs/resolve-sidebar.js
@@ -1,0 +1,6 @@
+const path = require('path')
+module.exports = require(path.resolve(
+  'src',
+  'gatsby-theme-iterative-docs',
+  'sidebar'
+))
