+++ plugins/gatsby-theme-iterative-docs/src/sidebar.d.ts
@@ -1,0 +1,9 @@
+declare interface NormalizedSidebarItem {
+  label: string
+  path: string
+  source: boolean | string
+}
+
+declare const sidebar: NormalizedSidebarItem[]
+
+export default sidebar
