+++ plugins/gatsby-theme-iterative-docs/src/sidebar.js
@@ -1,0 +1,1 @@
+module.exports = []
