+++ src/components/Documentation/WithJSX/index.tsx
@@ -4,7 +4,8 @@ import MarkdownMain from '../Markdown/Main'
 import RightPanel from '../RightPanel'
 
 import { IHeading, getGithubLink } from '../'
-import { getItemByPath } from '../../../utils/shared/sidebar'
+import { getItemByPath } from '../../../../plugins/gatsby-theme-iterative-docs/sidebar-helpers'
+import sidebar from '../../../../plugins/gatsby-theme-iterative-docs/src/sidebar'
 
 interface IWithJSXProps {
   path: string
@@ -12,7 +13,7 @@ interface IWithJSXProps {
 }
 
 const WithJSX: React.FC<IWithJSXProps> = ({ children, path, headings }) => {
-  const { source, prev, next, tutorials } = getItemByPath(path)
+  const { source, prev, next, tutorials } = getItemByPath(sidebar, path)
   const githubLink = getGithubLink(source)
 
   return (
