+++ src/components/Documentation/index.tsx
@@ -4,7 +4,8 @@ import { Node } from 'unist'
 import Markdown from './Markdown'
 import RightPanel from './RightPanel'
 
-import { getItemByPath } from '../../utils/shared/sidebar'
+import { getItemByPath } from '../../../plugins/gatsby-theme-iterative-docs/sidebar-helpers'
+import sidebar from '../../../plugins/gatsby-theme-iterative-docs/src/sidebar'
 
 export interface IHeading {
   slug: string
@@ -25,7 +26,7 @@ const Documentation: React.FC<IDocumentationProps> = ({
   path,
   headings
 }) => {
-  const { source, prev, next, tutorials } = getItemByPath(path)
+  const { source, prev, next, tutorials } = getItemByPath(sidebar, path)
   const githubLink = getGithubLink(source)
 
   return (
