+++ src/components/LayoutFooter/index.tsx
@@ -5,7 +5,8 @@ import LayoutWidthContainer from '../LayoutWidthContainer'
 import Link from '../Link'
 import SocialIcon, { ISocialIcon } from '../SocialIcon'
 import ShowOnly from '../ShowOnly'
-import { getFirstPage } from '../../utils/shared/sidebar'
+import { getFirstPage } from '../../../plugins/gatsby-theme-iterative-docs/sidebar-helpers'
+import sidebar from '../../../plugins/gatsby-theme-iterative-docs/src/sidebar'
 
 import { ReactComponent as LogoSVG } from '../../../static/img/dvc_icon-color--square_vector.svg'
 import { ReactComponent as GithubSVG } from '../SocialIcon/github.svg'
@@ -18,7 +19,7 @@ import { ReactComponent as MlemSVG } from '../../../static/img/mlem-icon.svg'
 
 import * as styles from './styles.module.css'
 
-const docsPage = getFirstPage()
+const docsPage = getFirstPage(sidebar)
 
 interface IFooterLinkData {
   href: string
