+++ src/gatsby-theme-iterative-docs/sidebar.js
@@ -1,0 +1,3 @@
+const normalizeSidebar = require('../../plugins/gatsby-theme-iterative-docs/normalize-sidebar')
+const sidebar = require('../../content/docs/sidebar.json')
+module.exports = normalizeSidebar(sidebar)
