+++ src/templates/doc-jsx.tsx
@@ -1,5 +1,6 @@
 import React from 'react'
-import { getItemByPath } from '../utils/shared/sidebar'
+import { getItemByPath } from '../../plugins/gatsby-theme-iterative-docs/sidebar-helpers'
+import sidebar from '../../plugins/gatsby-theme-iterative-docs/src/sidebar'
 
 import SEO from '../components/SEO'
 
@@ -19,7 +20,7 @@ const JSXDocPage: React.FC<IJSXDocPageProps> = ({
   slug,
   headings
 }) => {
-  const { label } = getItemByPath(slug)
+  const { label } = getItemByPath(sidebar, slug)
   return (
     <>
       <SEO title={title || label} description={description} />
