+++ src/templates/doc.tsx
@@ -1,7 +1,8 @@
 import React from 'react'
 import { graphql } from 'gatsby'
 import { Node } from 'unist'
-import { getItemByPath } from '../utils/shared/sidebar'
+import { getItemByPath } from '../../plugins/gatsby-theme-iterative-docs/sidebar-helpers'
+import sidebar from '../../plugins/gatsby-theme-iterative-docs/src/sidebar'
 
 import SEO from '../components/SEO'
 
@@ -35,7 +36,7 @@ const DocPage: React.FC<IDocPageProps> = ({
     }
   } = data
 
-  const { label } = getItemByPath(slug)
+  const { label } = getItemByPath(sidebar, slug)
 
   return (
     <>
